+++ ryoanji/src/ryoanji/interface/multipole_holder.cuh
@@ -59,7 +59,7 @@ public:
      * @param[inout] ugrav        potential per particle to add to, can be nullptr, on GPU
      * @param[inout] ax, ay, az   particle accelerations to add to, on GPU
      */
-    float compute(cstone::GroupView grp, const Tc* x, const Tc* y, const Tc* z, const Tm* m, const Th* h, Tc G,
+    double compute(cstone::GroupView grp, const Tc* x, const Tc* y, const Tc* z, const Tm* m, const Th* h, Tc G,
                   int numShells, const cstone::Box<Tc>& box, Ta* ugrav, Ta* ax, Ta* ay, Ta* az);
 
     /*! @brief compute accelerations on target particles
@@ -73,7 +73,7 @@ public:
      * @param[inout] ugrav           potential per particle to add to, can be nullptr, on GPU
      * @param[inout] ax, ay, az      particle accelerations to add to, on GPU
      */
-    float compute(cstone::GroupView grp, const Tc* xt, const Tc* yt, const Tc* zt, const Tm* mt, const Th* ht,
+    double compute(cstone::GroupView grp, const Tc* xt, const Tc* yt, const Tc* zt, const Tm* mt, const Th* ht,
                   const Tc* xs, const Tc* ys, const Tc* zs, const Tm* ms, const Th* hs, Tc G, int numShells,
                   const cstone::Box<Tc>& box, Ta* ugrav, Ta* ax, Ta* ay, Ta* az);
